# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AvatarSpeechRequest": {
        "properties": {
          "text": {
            "maxLength": 1200,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AvatarSpeechRequest",
        "type": "object"
      },
      "AvatarStatusResponse": {
        "properties": {
          "animation_model": {
            "title": "Animation Model",
            "type": "string"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "device": {
            "title": "Device",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "speech_model": {
            "title": "Speech Model",
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "available",
          "speech_model",
          "animation_model",
          "device"
        ],
        "title": "AvatarStatusResponse",
        "type": "object"
      },
      "Body_transcribe_v1_transcribe_post": {
        "properties": {
          "audio": {
            "contentMediaType": "application/octet-stream",
            "title": "Audio",
            "type": "string"
          }
        },
        "required": [
          "audio"
        ],
        "title": "Body_transcribe_v1_transcribe_post",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "FastAPI HTTP and validation error envelope.",
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              }
            ],
            "title": "Detail"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "error_code",
          "correlation_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvitationCreateRequest": {
        "properties": {
          "institution_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Institution Id",
            "type": "string"
          },
          "patient_sno": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Patient Sno",
            "type": "string"
          },
          "prefill": {
            "$ref": "#/components/schemas/InvitationPrefill"
          },
          "reg_sno": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Reg Sno",
            "type": "string"
          }
        },
        "required": [
          "institution_id",
          "patient_sno",
          "reg_sno"
        ],
        "title": "InvitationCreateRequest",
        "type": "object"
      },
      "InvitationExchangeRequest": {
        "properties": {
          "token": {
            "maxLength": 512,
            "minLength": 32,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "InvitationExchangeRequest",
        "type": "object"
      },
      "InvitationExchangeResponse": {
        "properties": {
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "expires_at"
        ],
        "title": "InvitationExchangeResponse",
        "type": "object"
      },
      "InvitationPrefill": {
        "description": "Minimum patient context accepted only from the authenticated UCC server.",
        "properties": {
          "allergies": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allergies"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "blood_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blood Type"
          },
          "current_medications": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Medications"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "medical_history": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medical History"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "surgical_history": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surgical History"
          }
        },
        "title": "InvitationPrefill",
        "type": "object"
      },
      "InvitationResponse": {
        "properties": {
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "invite_id": {
            "title": "Invite Id",
            "type": "string"
          },
          "public_url": {
            "title": "Public Url",
            "type": "string"
          },
          "status": {
            "const": "active",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "invite_id",
          "public_url",
          "expires_at",
          "status"
        ],
        "title": "InvitationResponse",
        "type": "object"
      },
...
      "PatientSessionResponse": {
        "properties": {
          "completed": {
            "default": false,
            "title": "Completed",
            "type": "boolean"
          },
          "consultation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consultation Id"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "interview_session_id": {
            "title": "Interview Session Id",
            "type": "string"
          },
          "status": {
            "const": "active",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "expires_at",
          "interview_session_id"
        ],
        "title": "PatientSessionResponse",
        "type": "object"
      },
...
      "TranscriptionResponse": {
        "properties": {
          "latency_seconds": {
            "minimum": 0.0,
            "title": "Latency Seconds",
            "type": "number"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "enum": [
              "breeze",
              "llm"
            ],
            "title": "Provider",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "provider",
          "model",
          "latency_seconds"
        ],
        "title": "TranscriptionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationIssue": {
        "properties": {
          "ctx": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctx"
          },
          "input": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "required": [
          "msg"
        ],
        "title": "ValidationIssue",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Versioned contract for the patient pre-consultation and physician workspace. Use `/v1` routes for new integrations; unversioned routes remain temporarily available as a compatibility layer.",
    "title": "AI 預問診系統 API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/avatar/speak": {
      "post": {
        "operationId": "avatar_speak_v1_avatar_speak_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ai_patient_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ai Patient Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarSpeechRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "video/mp4": {}
            },
            "description": "Locally generated talking-head MP4 video."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required or has expired."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation or domain validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An internal dependency or model operation failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required configured service is unavailable."
          }
        },
        "summary": "Render local CosyVoice3 speech as a MuseTalk avatar video",
        "tags": [
          "system"
        ]
      }
    },
    "/v1/avatar/status": {
      "get": {
        "operationId": "avatar_status_v1_avatar_status_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ai_patient_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ai Patient Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required or has expired."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check the private local avatar service",
        "tags": [
          "system"
        ]
      }
    },
...
    "/v1/invitations": {
      "post": {
        "operationId": "create_invitation_v1_invitations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller is not authorized for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation or domain validation failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required configured service is unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a single-use patient invitation from UCC",
        "tags": [
          "patient"
        ]
      }
    },
    "/v1/invitations/exchange": {
      "post": {
        "operationId": "exchange_invitation_v1_invitations_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationExchangeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request cannot be processed in its current state."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested consultation does not exist."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The submitted revision conflicts with the current revision."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation or domain validation failed."
          }
        },
        "summary": "Exchange an invitation token for an HttpOnly patient session",
        "tags": [
          "patient"
        ]
      }
    },
    "/v1/patient/session": {
      "get": {
        "operationId": "patient_session_v1_patient_session_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ai_patient_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ai Patient Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required or has expired."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore the authenticated patient interview session",
        "tags": [
          "patient"
        ]
      }
    },
    "/v1/transcribe": {
      "post": {
        "operationId": "transcribe_v1_transcribe_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ai_patient_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ai Patient Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_v1_transcribe_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request cannot be processed in its current state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The uploaded audio exceeds the accepted size limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation or domain validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An internal dependency or model operation failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required configured service is unavailable."
          }
        },
        "summary": "Transcribe a Traditional Chinese medical audio recording",
        "tags": [
          "system"
        ]
      }
    }
  },
...
}